Added ownership check for achievement management

+++ Systems/API/ZooExpoOrg.Api/Controllers/Achievements/AchievementController.cs
@@ -2,6 +2,7 @@
 using AutoMapper;
 using Azure.Core;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json.Linq;
 using ZooExpoOrg.Api.Controllers.Animals.Animals;
 using ZooExpoOrg.Api.Controllers.Clients;
 using ZooExpoOrg.Common.Exceptions;
@@ -9,6 +10,7 @@ using ZooExpoOrg.Services.Animals.Achievements;
 using ZooExpoOrg.Services.Animals.Animals;
 using ZooExpoOrg.Services.Clients;
 using ZooExpoOrg.Services.Logger;
+using ZooExpoOrg.Services.RightVerifier;
 
 namespace ZooExpoOrg.Api.Controllers.Achievements;
 
@@ -21,16 +23,19 @@ public class AchievementController : Controller
     private readonly IAppLogger logger;
     private readonly IAchievementService achievementService;
     private readonly IMapper mapper;
+    private readonly IRightVerifierService rightVerifier;
 
     public AchievementController(
         IAppLogger logger,
         IAchievementService achievementService,
-        IMapper mapper
+        IMapper mapper,
+        IRightVerifierService rightVerifier
         )
     {
         this.logger = logger;
         this.achievementService = achievementService;
         this.mapper = mapper;
+        this.rightVerifier = rightVerifier;
     }
 
     [HttpGet("owned/{ownerId:Guid}")]
@@ -71,24 +76,19 @@ public class AchievementController : Controller
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            Guid jwtClientId = await rightVerifier.GetClientId(jwtToken);
+            Guid requestClientId = await rightVerifier.GetClientIdByAnimalId(model.AnimalId);
+
+            if (requestClientId != jwtClientId)
+            {
+                return BadRequest("Access denied.");
+            }
+
             var achievements = await achievementService.Create(model);
 
             return Ok(mapper.Map<PresintationAchievementModel>(achievements));
-        }
-        catch (ProcessException e)
-        {
-            return NotFound(e.Message);
-        }
-    }
-
-    [HttpPut("{id:Guid}")]
-    public async Task<IActionResult> Update(Guid id, UpdateAchievementModel model)
-    {
-        try
-        {
-            await achievementService.Update(id, model);
-
-            return Ok();
         }
         catch (ProcessException e)
         {
@@ -101,6 +101,13 @@ public class AchievementController : Controller
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagAchievement(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             achievementService.Delete(id);
 
             return Ok();
